@@ -4,40 +4,43 @@ require 'spec_helper'
 describe "Search Results" do
   it "should have for an empty query" do
     search_for ''
-    number_of_results_from_page(page).should == 30
-    page.should have_xpath("//a[contains(@href, #{2007020969})]")
+    expect(number_of_results_from_page(page)).to eq(30)
+    expect(page).to have_xpath("//a[contains(@href, #{2007020969})]")
     search_for 'korea'
-    number_of_results_from_page(page).should == 4
+    expect(number_of_results_from_page(page)).to eq(4)
   end
 
   it "should find same result set with or without diacritcs" do
     search_for 'inmul'
-    number_of_results_from_page(page).should == 1
-    page.should have_xpath("//a[contains(@href, #{77826928})]")
+    expect(number_of_results_from_page(page)).to eq(1)
+    expect(page).to have_xpath("//a[contains(@href, #{77826928})]")
 
     search_for 'inmül'
-    number_of_results_from_page(page).should == 1
+    expect(number_of_results_from_page(page)).to eq(1)
   end
   it "should find same result set for a case-insensitive query " do
     search_for 'inmul'
-    number_of_results_from_page(page).should == 1
-    page.should have_xpath("//a[contains(@href, #{77826928})]")
+    expect(number_of_results_from_page(page)).to eq(1)
+    expect(page).to have_xpath("//a[contains(@href, #{77826928})]")
 
     search_for 'INMUL'
-    number_of_results_from_page(page).should == 1
+    expect(number_of_results_from_page(page)).to eq(1)
   end
 
   it "should order by relevancy" do
     search_for "Korea"
-    position_in_result_page(page, '77826928').should == 1
-    position_in_result_page(page, '94120425').should == 2
-    
+    expect(position_in_result_page(page, '77826928')).to eq(1)
+    expect(position_in_result_page(page, '94120425')).to eq(2)
+
   end
 
   it "should have an opensearch description document" do
     visit root_path
+    tmp_value = Capybara.ignore_hidden_elements
+    Capybara.ignore_hidden_elements = false
     page.should have_xpath("//link[contains(@rel, 'search')]")
     expect(page.find(:xpath, "//link[contains(@rel, 'search')]")[:href]).to eq "http://www.example.com/catalog/opensearch.xml"
+    Capybara.ignore_hidden_elements = tmp_value
   end
 
   it "should provide search hints if there are no results" do
@@ -46,6 +49,7 @@ describe "Search Results" do
   end
 
   it "should pass the current search id through", :js => true do
+    pending
     visit root_path
     fill_in "q", with: ''
     click_button 'search'
@@ -79,7 +83,7 @@ def position_in_result_page(page, id)
   end
   i.to_i
 end
-        
+
 def number_of_results_for_query(query)
   visit root_path
   fill_in "q", :with => query
@@ -88,5 +92,9 @@ def number_of_results_for_query(query)
 end
 
 def number_of_results_from_page(page)
-  page.find("meta[name=totalResults]")['content'].to_i rescue 0
+  tmp_value = Capybara.ignore_hidden_elements
+  Capybara.ignore_hidden_elements = false
+  val = page.find("meta[name=totalResults]")['content'].to_i rescue 0
+  Capybara.ignore_hidden_elements = tmp_value
+  val
 end
