@@ -13,41 +13,41 @@ describe SearchHistoryConstraintsHelper do
   end
 
   before(:each) do
-    helper.stub(:blacklight_config).and_return(@config)
+    allow(helper).to receive(:blacklight_config).and_return(@config)
   end
 
   describe "render_search_to_s_*" do
     describe "render_search_to_s_element" do
       it "should render basic element" do
         response = helper.render_search_to_s_element("key", "value")
-        response.should have_selector("span.constraint")  do |span|
-          span.should have_selector("span.filterName", :content => "key:")
-          span.should have_selector("span.filterValue", :content => "value")
+        expect(response).to have_selector("span.constraint")  do |span|
+          expect(span).to have_selector("span.filterName", :content => "key:")
+          expect(span).to have_selector("span.filterValue", :content => "value")
         end
-        response.html_safe?.should == true
+        expect(response.html_safe?).to eq(true)
       end
       it "should escape them that need escaping" do
         response = helper.render_search_to_s_element("key>", "value>")
-        response.should have_selector("span.constraint") do |span|          
-          span.should have_selector("span.filterName") do |s2|
+        expect(response).to have_selector("span.constraint") do |span|          
+          expect(span).to have_selector("span.filterName") do |s2|
             # Note: nokogiri's gettext will unescape the inner html
             # which seems to be what rspecs "contains" method calls on 
             # text nodes - thus the to_s inserted below.
-            s2.to_s.should match(/key&gt;:/)
+            expect(s2.to_s).to match(/key&gt;:/)
           end
-          span.should have_selector("span.filterValue") do |s3|            
-            s3.to_s.should match(/value&gt;/)
+          expect(span).to have_selector("span.filterValue") do |s3|            
+            expect(s3.to_s).to match(/value&gt;/)
           end
         end
-        response.html_safe?.should == true
+        expect(response.html_safe?).to eq(true)
       end
       it "should not escape with options set thus" do
         response = helper.render_search_to_s_element("key>", "value>", :escape_key => false, :escape_value => false)
-        response.should have_selector("span.constraint") do |span|
-          span.should have_selector("span.filterName", :content => "key>:")
-          span.should have_selector("span.filterValue", :content => "value>")
+        expect(response).to have_selector("span.constraint") do |span|
+          expect(span).to have_selector("span.filterName", :content => "key>:")
+          expect(span).to have_selector("span.filterValue", :content => "value>")
         end
-        response.html_safe?.should == true
+        expect(response.html_safe?).to eq(true)
       end
     end
 
@@ -56,15 +56,15 @@ describe SearchHistoryConstraintsHelper do
         @params = {:q => "history", :f => {"some_facet" => ["value1", "value1"],  "other_facet" => ["other1"]}}        
       end
       it "should call lesser methods" do
-        helper.stub(:blacklight_config).and_return(@config)
-        helper.stub(:default_search_field).and_return(Blacklight::Configuration::SearchField.new(:key => 'default_search_field', :display_label => 'Default'))
-        helper.stub(:label_for_search_field).with(nil).and_return('')
+        allow(helper).to receive(:blacklight_config).and_return(@config)
+        allow(helper).to receive(:default_search_field).and_return(Blacklight::Configuration::SearchField.new(:key => 'default_search_field', :display_label => 'Default'))
+        allow(helper).to receive(:label_for_search_field).with(nil).and_return('')
         # API hooks expect this to be so
         response = helper.render_search_to_s(@params)
 
-        response.should include( helper.render_search_to_s_q(@params))
-        response.should include( helper.render_search_to_s_filters(@params))
-        response.html_safe?.should == true
+        expect(response).to include( helper.render_search_to_s_q(@params))
+        expect(response).to include( helper.render_search_to_s_filters(@params))
+        expect(response.html_safe?).to eq(true)
       end
     end
     
