@@ -6,16 +6,16 @@ describe 'Blacklight::Utils' do
     it "should provide hash-like accessors for OpenStruct data" do
       a = Blacklight::OpenStructWithHashAccess.new :foo => :bar, :baz => 1
 
-      a[:foo].should == :bar
-      a[:baz].should == 1
-      a[:asdf].should be_nil
+      expect(a[:foo]).to eq(:bar)
+      expect(a[:baz]).to eq(1)
+      expect(a[:asdf]).to be_nil
     end
 
     it "should provide hash-like writers for OpenStruct data" do
       a = Blacklight::OpenStructWithHashAccess.new :foo => :bar, :baz => 1
 
       a[:asdf] = 'qwerty'
-      a.asdf.should == 'qwerty'
+      expect(a.asdf).to eq('qwerty')
 
     end
     
@@ -23,12 +23,12 @@ describe 'Blacklight::Utils' do
       h = Blacklight::OpenStructWithHashAccess.new
       
       h["string"] = "value"
-      h[:string].should == "value"
-      h.string.should == "value"
+      expect(h[:string]).to eq("value")
+      expect(h.string).to eq("value")
       
       h[:symbol] = "value"
-      h["symbol"].should == "value"
-      h.symbol.should == "value"      
+      expect(h["symbol"]).to eq("value")
+      expect(h.symbol).to eq("value")      
     end
 
     describe "internal hash table" do
@@ -39,12 +39,12 @@ describe 'Blacklight::Utils' do
       end
 
       it "should expose the internal hash table" do
-        @h.to_h.should be_a_kind_of(Hash)
-        @h.to_h[:a].should == 1
+        expect(@h.to_h).to be_a_kind_of(Hash)
+        expect(@h.to_h[:a]).to eq(1)
       end
 
       it "should expose keys" do
-        @h.keys.should include(:a, :b)
+        expect(@h.keys).to include(:a, :b)
       end
     end
 
