@@ -3,11 +3,13 @@ require 'spec_helper'
 describe "Facets" do
   it "should work without a search term" do
     visit root_path
-    click_link "Tibetan"
-    within ("#sortAndPerPage") do
+    within 'div.blacklight-language_facet' do
+      click_link "Tibetan"
+    end
+    within "#sortAndPerPage" do
       expect(page).to have_content "1 - 6 of 6"
     end
-    within(".blacklight-language_facet") do 
+    within("div.blacklight-language_facet") do
       expect(page).to have_selector("span.selected", :text => "Tibetan 6")
     end
 
@@ -15,10 +17,10 @@ describe "Facets" do
     within ("#sortAndPerPage") do
       expect(page).to have_content "1 - 2 of 2"
     end
-    within(".blacklight-language_facet") do 
+    within("div.blacklight-language_facet") do
       expect(page).to have_selector("span.selected", :text => "Tibetan 2")
     end
-    within(".blacklight-subject_geo_facet") do 
+    within(".blacklight-subject_geo_facet") do
       expect(page).to have_selector("span.selected", :text => "India 2")
     end
   end
@@ -31,11 +33,13 @@ describe "Facets" do
       expect(page).to have_content "1 - 9 of 9"
     end
 
-    click_link "Tibetan"
+    within "div.blacklight-language_facet" do
+      click_link "Tibetan"
+    end
     within ("#sortAndPerPage") do
       expect(page).to have_content "1 - 2 of 2"
     end
-    within(".blacklight-language_facet") do 
+    within("div.blacklight-language_facet") do
       expect(page).to have_selector("span.selected", :text => "Tibetan 2")
     end
     within "#appliedParams" do
@@ -48,21 +52,23 @@ describe "Facets" do
     within ("#sortAndPerPage") do
       expect(page).to have_content "1 entry found"
     end
-    within(".blacklight-language_facet") do 
+    within("div.blacklight-language_facet") do
       expect(page).to have_selector("span.selected", :text => "Tibetan 1")
     end
-    within(".blacklight-pub_date") do 
+    within(".blacklight-pub_date") do
       expect(page).to have_selector("span.selected", :text => "2004 1")
     end
   end
 
   it "should allow removing filters" do
     visit root_path
-    click_link "Tibetan"
-    within(".blacklight-language_facet") do 
+    within "div.blacklight-language_facet" do
+      click_link "Tibetan"
+    end
+    within("div.blacklight-language_facet") do
       expect(page).to have_selector("span.selected", :text => "Tibetan 6")
     end
-    within(".blacklight-language_facet") do 
+    within("div.blacklight-language_facet") do
       click_link 'remove'
     end
     expect(page).to_not have_link 'remove'
@@ -73,26 +79,28 @@ describe "Facets" do
     visit root_path
     fill_in "q", with: 'history'
     click_button 'search'
-    click_link 'Tibetan'
-    within(".blacklight-language_facet") do 
+    within "div.blacklight-language_facet" do
+      click_link "Tibetan"
+    end
+    within("div.blacklight-language_facet") do
       expect(page).to have_selector("span.selected", :text => "Tibetan 2")
     end
 
     click_link '2004'
 
-    within(".blacklight-language_facet") do 
+    within("div.blacklight-language_facet") do
       expect(page).to have_selector("span.selected", :text => "Tibetan 1")
     end
-    within(".blacklight-pub_date") do 
+    within(".blacklight-pub_date") do
       expect(page).to have_selector("span.selected", :text => "2004 1")
     end
     fill_in "q", with: 'china'
     click_button 'search'
 
-    within(".blacklight-language_facet") do 
+    within("div.blacklight-language_facet") do
       expect(page).to have_selector("span.selected", :text => "Tibetan 1")
     end
-    within(".blacklight-pub_date") do 
+    within(".blacklight-pub_date") do
       expect(page).to have_selector("span.selected", :text => "2004 1")
     end
   end
@@ -101,12 +109,14 @@ describe "Facets" do
     visit root_path
     fill_in "q", with: 'history'
     click_button 'search'
-    click_link 'Tibetan'
-    within(".blacklight-language_facet") do 
+    within("div.blacklight-language_facet") do
+      click_link 'Tibetan'
+    end
+    within("div.blacklight-language_facet") do
       expect(page).to have_selector("span.selected", :text => "Tibetan 2")
     end
     click_link 'title'
-    within(".blacklight-language_facet") do 
+    within("div.blacklight-language_facet") do
       expect(page).to have_selector("span.selected", :text => "Tibetan 2")
     end
     within "#appliedParams" do
@@ -119,14 +129,17 @@ describe "Facets" do
     visit root_path
     fill_in "q", with: 'history'
     click_button 'search'
-    click_link 'Tibetan'
-    within(".blacklight-language_facet") do 
+
+    within("div.blacklight-language_facet") do
+      click_link 'Tibetan'
+    end
+    within("div.blacklight-language_facet") do
       expect(page).to have_selector("span.selected", :text => "Tibetan 2")
     end
     within '#per_page-dropdown' do
       click_link '20'
     end
-    within(".blacklight-language_facet") do 
+    within("div.blacklight-language_facet") do
       expect(page).to have_selector("span.selected", :text => "Tibetan 2")
     end
     within "#appliedParams" do
