@@ -3,7 +3,9 @@ require 'spec_helper'
 describe "Search Sort" do
   it "should sort on facet results with no search terms" do
     visit root_path
-    click_link 'English'
+    within("div.blacklight-language_facet") do
+      click_link 'English'
+    end
     expect(page).to have_content 'Sort by relevance'
     click_link 'title'
     expect(page).to have_content 'Sort by title'
