@@ -14,9 +14,9 @@ module Features
       email = "#{login}@#{login}.com"
       user = User.create(:email => email, :password => "password", :password_confirmation => "password")
       visit new_user_session_path
-      fill_in("user_email", :with => email) 
-      fill_in("user_password", :with => "password") 
-      click_button("Sign in")
+      fill_in("user_email", :with => email)
+      fill_in("user_password", :with => "password")
+      click_button("Log in")
     end
   end
 end
