@@ -18,9 +18,9 @@ describe "catalog/_thumbnail_default" do
 
   before do
     assign :response, double(:params => {})
-    view.stub(:render_grouped_response?).and_return false
-    view.stub(:blacklight_config).and_return(blacklight_config)
-    view.stub(:current_search_session).and_return nil
+    allow(view).to receive(:render_grouped_response?).and_return false
+    allow(view).to receive(:blacklight_config).and_return(blacklight_config)
+    allow(view).to receive(:current_search_session).and_return nil
   end
 
   it "should render the thumbnail if the document has one" do
