@@ -2,7 +2,7 @@ require 'spec_helper'
 
 describe "catalog/index.html.erb" do
   it "should render the sidebar and content panes" do
-    view.stub(:blacklight_config).and_return(Blacklight::Configuration.new)
+    allow(view).to receive(:blacklight_config).and_return(Blacklight::Configuration.new)
     render
     expect(rendered).to match /id="sidebar"/
     expect(rendered).to match /id="content"/
@@ -11,7 +11,7 @@ describe "catalog/index.html.erb" do
   it "should render the search_sidebar partial " do
     stub_template "catalog/_search_sidebar.html.erb" => "sidebar_content"
 
-    view.stub(:blacklight_config).and_return(Blacklight::Configuration.new)
+    allow(view).to receive(:blacklight_config).and_return(Blacklight::Configuration.new)
     render
     expect(rendered).to match /sidebar_content/
   end
@@ -20,10 +20,10 @@ describe "catalog/index.html.erb" do
     stub_template "catalog/_results_pagination.html.erb" => ""
     stub_template "catalog/_search_header.html.erb" => "header_content"
 
-    view.stub(:blacklight_config).and_return(Blacklight::Configuration.new)
-    view.stub(:has_search_parameters?).and_return(true)
-    view.stub(:extra_head_content).and_return([])
-    view.stub(:render_opensearch_response_metadata).and_return("")
+    allow(view).to receive(:blacklight_config).and_return(Blacklight::Configuration.new)
+    allow(view).to receive(:has_search_parameters?).and_return(true)
+    allow(view).to receive(:extra_head_content).and_return([])
+    allow(view).to receive(:render_opensearch_response_metadata).and_return("")
     assign(:response, double(:empty? => true))
     render
     expect(rendered).to match /header_content/
